were added some chnges incuding marks, comments and also working of app was established
