+++ AnimationsApp/Delegates/SceneDelegate.swift
@@ -17,9 +17,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-//        let navigationController = UINavigationController(rootViewController: ProfileViewController())
-//        navigationController.navigationBar.prefersLargeTitles = true
-//        window?.rootViewController = navigationController
         window?.rootViewController = ProfileViewController()
         window?.makeKeyAndVisible()
     }
